Graphql Queries
Updated gatsby-node.js to fetch the previous and next post in relations to the current post being viewed by the user. This is then accessed via a link at the bottom of each post. Adjustments to media queries were also made for smaller phones in portrait view such as the iPhone 5/SE